*: improve planner/core to make test faster
